Added GUIController and example button
When TouchHandler needs a confirmation button, it flags GUIController
appropriately, and asks GUIController for the player's response.

When the GUIController is flagged, it begins animation until the
"buttons" are in the right place. Then it displays the buttons and will
display them every update until a player gives a response.

The response is set to 0 in GUIController if there hasn't been a
response this update, since TouchHandler will only accept a response of
'1' or '2'.